+++ school-medical-management/src/main/java/com/swp391/school_medical_management/modules/users/controllers/ParentController.java
@@ -231,7 +231,7 @@ public class ParentController {
     }
 
     @GetMapping("/health-check-result/{healCheckResultId}")
-    public ResponseEntity<HealthCheckResultDTO> getHealthCheckResultByResultID(@PathVariable Long healCheckResultId) {
+    public ResponseEntity<HealthCheckResultDTO> getHealthCheckResultByResultID(@PathVariable int healCheckResultId) {
         HealthCheckResultDTO healthCheckResultDTO = nurseService.getHealthCheckResult(healCheckResultId);
         return ResponseEntity.ok(healthCheckResultDTO);
     }
